+++ HISTORY.md
@@ -2,12 +2,14 @@
 
 ## in development:
 * **New Features:**
+    * new "rgba" option in "(aspect) AT (location)" sensing reporter, returns a 4-item list of values from 0-255 (same is pixels from a costume)
 * **Notable Changes:**
 * **Notable Fixes:**
 * **Translation Updates:**
 
 ### 2019-08-07
 * new dev version
+* blocks, threads: added "rgba" option to (aspect) AT (location) reporter in the sensing category
 
 ## v5.0.9
 * **New Feature:**